requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
alembic==1.13.1
qrcode==7.4.2
Pillow==10.3.0
python-multipart==0.0.9
python-dotenv